updated tests with new names
