+++ tests/dummy/app/templates/index.hbs
@@ -1,101 +1,101 @@
-{{moment now}}
+{{moment-format now}}
 
 <code>
-  \{{moment now}}
+  \{{moment-format now}}
 </code>
 
 <hr>
 
-{{moment now 'LLLL' }}
+{{moment-format now 'LLLL' }}
 
 <code>
-  \{{moment now 'LLLL'}}
+  \{{moment-format now 'LLLL'}}
 </code>
 
 <hr>
 
-{{moment usIndependenceDay 'LLLL' }}
+{{moment-format usIndependenceDay 'LLLL' }}
 
 <code>
-  \{{moment usIndependenceDay 'LLLL'}}
+  \{{moment-format usIndependenceDay 'LLLL'}}
 </code>
 
 <hr>
 
-{{moment usIndependenceDay 'MMM DD, YYYY' }}
+{{moment-format usIndependenceDay 'MMM DD, YYYY' }}
 
 <code>
-  \{{moment usIndependenceDay 'MMM DD, YYYY'}}
-</code>
-
-
-<hr>
-
-{{moment now 'LL' 'LLLL'}}
-
-<code>
-  \{{moment now 'LL' 'LLLL'}}
+  \{{moment-format usIndependenceDay 'MMM DD, YYYY'}}
 </code>
 
 
 <hr>
 
-{{ago now}}
+{{moment-format now 'LL' 'LLLL'}}
 
 <code>
-  \{{ago now}}
+  \{{moment-format now 'LL' 'LLLL'}}
 </code>
 
 
 <hr>
 
-{{ago lastHour}}
+{{moment-ago now}}
 
 <code>
-  \{{ago lastHour}}
+  \{{moment-ago now}}
+</code>
+
+
+<hr>
+
+{{moment-ago lastHour}}
+
+<code>
+  \{{moment-ago lastHour}}
 </code>
 
 <hr>
 
-{{ago usIndependenceDay}}
+{{moment-ago usIndependenceDay}}
 
 <code>
-  \{{ago usIndependenceDay}}
+  \{{moment-ago usIndependenceDay}}
 </code>
 
 <hr>
 
-{{duration 5000}}
+{{moment-duration 5000}}
 <code>
-  \{{duration 5000}}
+  \{{moment-duration 5000}}
 </code>
 
 <hr>
 
-{{duration 600000}}
+{{moment-duration 600000}}
 <code>
-  \{{duration 600000}}
+  \{{moment-duration 600000}}
 </code>
 
 <hr>
 
-{{duration 86400000}}
+{{moment-duration 86400000}}
 <code>
-  \{{duration 86400000}}
+  \{{moment-duration 86400000}}
 </code>
 
 <hr>
 
-{{duration '7.23:59:59'}}
+{{moment-duration '7.23:59:59'}}
 <code>
-  \{{duration '7.23:59:59'}}
+  \{{moment-duration '7.23:59:59'}}
 </code>
 
 <hr>
 
-{{duration 210 'days'}}
+{{moment-duration 210 'days'}}
 <code>
-  \{{duration 210 'days'}}
+  \{{moment-duration 210 'days'}}
 </code>
 
 <hr>
